Extract implementations of parallel_utils.hpp functions into a separate file to avoid multiple-definition linker errors
