+++ CMakeLists.txt
@@ -328,6 +328,7 @@ add_subdirectory("src/geojson")
 add_subdirectory("src/bmi")
 add_subdirectory("src/realizations/catchment")
 add_subdirectory("src/forcing")
+add_subdirectory("src/utilities")
 add_subdirectory("src/utilities/mdarray")
 add_subdirectory("src/utilities/mdframe")
 add_subdirectory("src/utilities/logging")
@@ -347,6 +348,7 @@ target_link_libraries(ngen
         NGen::logging
         NGen::bmi_protocols
         NGen::output
+        NGen::parallel
 )
 
 if(NGEN_WITH_SQLITE)
